% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wal.R
\name{apply.palette.to.rawdata}
\alias{apply.palette.to.rawdata}
\title{Apply a palette to index data to create a 2D image.}
\usage{
apply.palette.to.rawdata(raw_data, apply_palette, img_width, img_height)
}
\arguments{
\item{raw_data}{integer vector of pixel data, each entry represents an index into the palette.}

\item{apply_palette}{integer matrix, the palette.}

\item{img_width}{integer, the width of the image to create.}

\item{img_height}{integer, the height of the image to create.}
}
\description{
Apply a palette to index data to create a 2D image.
}
\keyword{internal}
