% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wal.R
\name{check.palette}
\alias{check.palette}
\title{Check palette, stop on invalid data.}
\usage{
check.palette(pal)
}
\arguments{
\item{pal}{a palette, i.e., a 256 x 3 integer matrix, with values in range 0..255L.}
}
\description{
Check palette, stop on invalid data.
}
\keyword{internal}
