% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_wal.R
\name{expand.rawdata.to.mipmaps}
\alias{expand.rawdata.to.mipmaps}
\title{Given the pixel data for the largest image, generate the full data for all mipmaps.}
\usage{
expand.rawdata.to.mipmaps(raw_data_mip_level0, width, height, byrow = TRUE)
}
\arguments{
\item{raw_data_mip_level0}{integer vector or matrix, the image data for the largest mipmap.}

\item{width}{integer, width of image for mip level 0}

\item{height}{integer, width of image for mip level 0}
}
\description{
Given the pixel data for the largest image, generate the full data for all mipmaps.
}
\keyword{internal}
