% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_wal.R
\name{get.mipmap.data.lengths}
\alias{get.mipmap.data.lengths}
\title{Compute length of mipmaps in bytes from width and height of largest image (mipmap0).}
\usage{
get.mipmap.data.lengths(mm0_width, mm0_height)
}
\arguments{
\item{mm0_width}{integer, width of mipmap 0}

\item{mm0_height}{integer, height of mipmap 0}
}
\value{
integer vector of length 4, the lengths.
}
\description{
Compute length of mipmaps in bytes from width and height of largest image (mipmap0).
}
\keyword{internal}
