% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_wal.R
\name{get.mipmap.data.offsets}
\alias{get.mipmap.data.offsets}
\title{Get mipmap offsets for WAL header, based on mipmap sizes and start offset.}
\usage{
get.mipmap.data.offsets(mm0_width, mm0_height, start_at = 100L)
}
\arguments{
\item{mm0_width}{integer, width of mipmap 0}

\item{mm0_height}{integer, height of mipmap 0}

\item{start_at}{integer, the offset at which the data starts in the file. Must be 100L for WAL format.}
}
\value{
integer vector of length 4, the offsets.
}
\description{
Get mipmap offsets for WAL header, based on mipmap sizes and start offset.
}
\keyword{internal}
