% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wad.R
\name{get.wadentry.type.strings}
\alias{get.wadentry.type.strings}
\title{Translate wad directory entry types from the integer to the string representation.}
\usage{
get.wadentry.type.strings(wadentry_type_int)
}
\arguments{
\item{wadentry_type_int}{integer, WAD entry type code}
}
\value{
type string
}
\description{
Translate wad directory entry types from the integer to the string representation.
}
\keyword{internal}
