% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wal.R
\name{get.wal.mipmap.heights}
\alias{get.wal.mipmap.heights}
\title{Compute widths of the 4 mipimap levels from base width.}
\usage{
get.wal.mipmap.heights(height_mm0)
}
\arguments{
\item{height_mm}{interger, the base mipmap height.}
}
\value{
integer vector of length 4, the mipmap heights.
}
\description{
Compute widths of the 4 mipimap levels from base width.
}
\keyword{internal}
