% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wal.R
\name{get.wal.mipmap.widths}
\alias{get.wal.mipmap.widths}
\title{Compute widths of the 4 mipimap levels from base width.}
\usage{
get.wal.mipmap.widths(width_mm0)
}
\arguments{
\item{width_mm}{integer, the base mipmap width.}
}
\value{
integer vector of length 4, the mipmap widths.
}
\description{
Compute widths of the 4 mipimap levels from base width.
}
\keyword{internal}
