% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_wal.R
\name{half.image}
\alias{half.image}
\title{Reduce image size by 2 along both axes by dropping pixels.}
\usage{
half.image(image_data, byrow = TRUE)
}
\arguments{
\item{image_data}{integer matrix, a 1-channel image.}
}
\description{
Reduce image size by 2 along both axes by dropping pixels.
}
\keyword{internal}
