% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pak.R
\name{pak.create}
\alias{pak.create}
\title{Create a PAK file from the contents of a directory.}
\usage{
pak.create(out_pakfile, contents_dir)
}
\arguments{
\item{out_pakfile}{character string, the full path and PAK filename that should be created. If only a filename is given, it will be created in the current working directory.}

\item{contents_dir}{character string, a directory. The contents, without the directory itself, will become the PAK file contents. Must exist and be readable.}
}
\description{
Create a PAK file from the contents of a directory.
}
\examples{
\dontrun{
    pd = "~/data/q1_pak0_extracted";
    pak.create("~/mypak0.pak", pd);
}

}
