% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pak.R
\name{pak.extract}
\alias{pak.extract}
\title{Extract PAK contents into existing directory.}
\usage{
pak.extract(pak_filepath, outdir = getwd())
}
\arguments{
\item{pak_filepath}{character string, path to input PAK file.}

\item{outdir}{character string, the output directory in which the files should be created. Must be writeable. The sub directories and filenames are derived from the data in the WAD.}
}
\description{
Extract PAK contents into existing directory.
}
\note{
PAK files can contain a directory structure, and new subdirectories will be created under \code{outdir} as needed to preserve it.
}
