% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pak.R
\name{pak.header.create}
\alias{pak.header.create}
\title{Generate PAK header from filetable.}
\usage{
pak.header.create(filetable, filetable_at_end = TRUE)
}
\arguments{
\item{filetable}{data.frame, the PAK contents filetable.}

\item{filetable_at_end}{logical, whether to place the filetable at the end of the file, after all file data. If FALSE, it will be places after the PAK header, before the file data. Should not matter, leave this alone if in doubt.}
}
\value{
named list, the PAK stub. The only entry is the 'header' list.
}
\description{
Generate PAK header from filetable.
}
\keyword{internal}
