% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pak.R
\name{pak.write}
\alias{pak.write}
\title{Write a PAK file based on the filetable.}
\usage{
pak.write(out_pakfile, contents_dir, filetable, filetable_at_end = TRUE)
}
\arguments{
\item{out_pakfile}{character string, the full path and PAK filename that should be created. If only a filename is given, it will be created in the current working directory.}

\item{contents_dir}{character string, a directory. The contents, without the directory itself, will become the PAK file contents. Must exist and be readable.}

\item{filetable}{data.frame, the PAK contents filetable.}

\item{filetable_at_end}{logical, whether to place the filetable at the end of the file, after all file data. If FALSE, it will be places after the PAK header, before the file data. Should not matter, leave this alone if in doubt.}
}
\description{
Write a PAK file based on the filetable.
}
\keyword{internal}
