% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{pal_q2}
\alias{pal_q2}
\title{Get Q2 palette.}
\usage{
pal_q2()
}
\value{
256 x 3 integer matrix, representing the RGB color values for an index into the palette.
}
\description{
Get Q2 palette.
}
\examples{
   pal = pal_q2();
   dim(pal);

}
