% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{palette_q1}
\alias{palette_q1}
\title{Get Q1 palette.}
\usage{
palette_q1()
}
\value{
256 x 3 integer matrix, representing the RGB color values for an index into the palette.
}
\description{
Get Q1 palette.
}
