% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wal.R
\name{plot.wal}
\alias{plot.wal}
\title{S3 plot function for wal image.}
\usage{
\method{plot}{wal}(x, ...)
}
\arguments{
\item{x}{a wal instance.}

\item{...}{extra args, not used.}
}
\description{
S3 plot function for wal image.
}
