% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wad.R
\name{print.wad}
\alias{print.wad}
\title{S3 print function for WAD}
\usage{
\method{print}{wad}(x, ...)
}
\arguments{
\item{x}{wad instance}

\item{...}{extra arguments, ignored}
}
\description{
S3 print function for WAD
}
