% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quake_archive.R
\name{qarchive.extract}
\alias{qarchive.extract}
\title{Extract any of the supported Quake archives.}
\usage{
qarchive.extract(filepath, outdir, format = "auto", do_pre_checks = TRUE)
}
\arguments{
\item{filepath}{character string, path to existing and readable file in PAK or WAD2 format.}

\item{outdir}{character string, path to an existing and writeable output directory into which to extract the archive.}

\item{format}{character string, of one 'auto' to detect from filename, 'QARCHIVE_TYPE_WAD' for WAD2, or 'QARCHIVE_TYPE_PAK' for PACK.}

\item{do_pre_checks}{logical, whether to perform extra sanity checks on the other parameters.}
}
\description{
Extract any of the supported Quake archives.
}
