% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quake_archive.R
\name{qarchive.type.from.filename}
\alias{qarchive.type.from.filename}
\title{Determine archive type from file name extension.}
\usage{
qarchive.type.from.filename(filepath)
}
\arguments{
\item{filepath}{character string, path to existing and readable file in PAK or WAD2 format.}
}
\value{
character string, one of 'QARCHIVE_TYPE_WAD' or 'QARCHIVE_TYPE_PAK'.
}
\description{
Determine archive type from file name extension.
}
\keyword{internal}
