% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_common.R
\name{quakemodel.to.fs.surface}
\alias{quakemodel.to.fs.surface}
\title{Convert Quake Model to 'fs.surface' instance.}
\usage{
quakemodel.to.fs.surface(quakemodel, frame_idx = 1L)
}
\arguments{
\item{quakemodel}{an instance of \code{quakemodel_mdl} or \code{quakemodel_md2}.}

\item{frame_idx}{integer, the frame to export. Quake models may contain animations made up of several frames. The mesh connectivity is unaltered between frames, but the vertex positions differ.}
}
\value{
\code{fs.surface} mesh instance, as used by the \code{freesurferformats} package.
}
\description{
Convert Quake Model to 'fs.surface' instance.
}
