% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{read.lmp}
\alias{read.lmp}
\title{Read binary lump, or 'lmp' files.}
\usage{
read.lmp(filepath, dlength = 768L)
}
\arguments{
\item{filepath}{character string, path to the input file.}

\item{dlength}{the expected data length, in bytes.}
}
\value{
vector of dlength unsigned integers in range 0..255.
}
\description{
Read binary lump, or 'lmp' files.
}
\keyword{internal}
