% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pak.R
\name{read.pak}
\alias{read.pak}
\title{Read Quake PAK archive.}
\usage{
read.pak(filepath)
}
\arguments{
\item{filepath}{character string, path to the file including extension.}
}
\value{
a 'pak' instance.
}
\description{
Read Quake PAK archive.
}
\examples{
\dontrun{
   pakf = '~/.steam/steam/steamapps/common/Quake/Id1/PAK0.PAK';
   pak = read.pak(pakf);
}

}
