% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wad.R
\name{read.wad}
\alias{read.wad}
\title{Read Quake WAD file.}
\usage{
read.wad(filepath)
}
\arguments{
\item{filepath}{character string, path to the file.}
}
\value{
a wad instance, can be used to extract data or list contents.
}
\description{
Read Quake WAD file.
}
\examples{
\dontrun{
   wadf = '~/knave.wad';
   wad = read.wad(wadf);
   wad.contents(wad);
}

}
