% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wal.R
\name{read.wal}
\alias{read.wal}
\title{Read bitmap file in WAL format.}
\usage{
read.wal(filepath, hdr = TRUE, hdr_only = FALSE)
}
\arguments{
\item{filepath}{character string, path to the file including extension}

\item{hdr}{logical, whether to return full list with header}

\item{hdr_only}{logical, whether to read only the header}
}
\value{
integer pixel matrix, each pixel value is in range 0-255 and refers to an index in a palette. The palette is NOT included in the file, so you will need to define one or get it from elsewhere to see the final image.
}
\description{
Read bitmap file in WAL format.
}
\examples{
\dontrun{
   walf = '~/data/q2_pak0_extracted/textures/e1u2/basic1_7.wal';
   wal = read.wal(walf);
   #plot(imager::as.cimg(wal$image))
}

}
