% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wad.R
\name{save.filepart}
\alias{save.filepart}
\title{Read part of binary file and save as new file.}
\usage{
save.filepart(infile, read_from, read_len, outfile)
}
\arguments{
\item{infile}{for input file, part of it gets read.}

\item{read_from}{integer, index at which to start reading, from start of file. Used to \code{seek} to the position.}

\item{read_len}{integer, the number of bytes to read.}

\item{outfile}{character string, the output filename.}
}
\description{
Read part of binary file and save as new file.
}
\keyword{internal}
