% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wad.R
\name{wad.contents}
\alias{wad.contents}
\title{List WAD file contents.}
\usage{
wad.contents(wad)
}
\arguments{
\item{wad}{a wad instance, see \code{read.wad}}
}
\value{
data.frame, the files inside the wad.
}
\description{
List WAD file contents.
}
