% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wad.R
\name{wad.texname.clean}
\alias{wad.texname.clean}
\title{Replace special chars in texture names to turn it into a valid filename.}
\usage{
wad.texname.clean(texnames)
}
\arguments{
\item{texnames}{character string, texture names from a WAD file. The textures may contain the special characters '*' and '+', which are used to indicate sequences (textures that change on an event, like a pressed button turning from red to green) and other things.}
}
\value{
character strings usable as filenames.
}
\description{
Replace special chars in texture names to turn it into a valid filename.
}
\keyword{internal}
