% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wad.R
\name{wad_dir.fileext.mapping}
\alias{wad_dir.fileext.mapping}
\title{Get file extensions for WAD dir entry type strings.}
\usage{
wad_dir.fileext.mapping()
}
\value{
named list, which maps \code{wad_dir.types.string}s to file extensions. Afaik, there are not standard file extensions for these file types, and I made the ones used here up.
}
\description{
Get file extensions for WAD dir entry type strings.
}
\seealso{
wad_dir.types.string
}
\keyword{internal}
