% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wad.R
\name{wad_dir.types.int}
\alias{wad_dir.types.int}
\title{Get integers representing WAD dir entry types.}
\usage{
wad_dir.types.int()
}
\description{
Get integers representing WAD dir entry types.
}
\seealso{
wad_dir.types.string
}
\keyword{internal}
