% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wad.R
\name{wad_dir.types.string}
\alias{wad_dir.types.string}
\title{Get strings describing WAD dir entry types.}
\usage{
wad_dir.types.string()
}
\description{
Get strings describing WAD dir entry types.
}
\seealso{
wad_dir.types.int
}
\keyword{internal}
