% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_to_wal.R
\name{wal.template}
\alias{wal.template}
\title{Generate a WAL structure template.}
\usage{
wal.template()
}
\description{
Generates a WAL instance that can be modified and filled with new data. The template represents a black 32x32 image (if the palette used to display it adhers to the convention that the first color is black). The indices used are 1-based (black is at index 1, not 0).
}
\keyword{internal}
