% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_wal.R
\name{writeWAL}
\alias{writeWAL}
\title{Write WAL instance to bitmap file in WAL format.}
\usage{
writeWAL(filepath, wal)
}
\arguments{
\item{filepath}{character string, path to the file including extension}

\item{wal}{a wal instance. Note that 1 will be substracted from the data when it is written, as indices are stored 0-based in the file.}
}
\description{
Write WAL instance to bitmap file in WAL format.
}
\examples{
\dontrun{
   walf = '~/data/q2_pak0_extracted/textures/e1u2/basic1_7.wal';
   wal = read.wal(walf);
   writeWAL(tempfile(fileext = ".wal"), wal);
}

}
