% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wal.R
\name{read.wal}
\alias{read.wal}
\title{Read bitmap file in WAL format.}
\usage{
read.wal(filepath, hdr = TRUE, hdr_only = FALSE, num_mip_maps = 4L)
}
\arguments{
\item{filepath}{character string, path to the file including extension}

\item{hdr}{logical, whether to return full list with header}

\item{hdr_only}{logical, whether to read only the header}

\item{num_mip_maps}{integer, the number of mipmap levels stored in the file. Leave this alone unless you know what you are doing.}
}
\value{
array with color data, or wal instance (named list) if `hdr` is `TRUE`
}
\description{
Read bitmap file in WAL format.
}
\examples{
\dontrun{
   walf = '~/data/q2_pak0_extracted/textures/e1u2/basic1_7.wal';
   wal = read.wal(walf);
   plot(wal);
   # show palette:
   plot(1:256, col=rgb(wal$palette, maxColorValue = 255));
}

}
