% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wad.R
\name{wad.extract}
\alias{wad.extract}
\title{Extract WAD contents into existing directory.}
\usage{
wad.extract(wad_filepath, outdir = getwd())
}
\arguments{
\item{wad_filepath}{character string, path to input WAD file.}

\item{outdir}{character string, the output directory in which the files should be created. The filenames are derived from the data in the WAD.}
}
\description{
Extract WAD contents into existing directory.
}
\note{
One can read extracted textures with \code{read.quake1miptex()}.
}
